.history/requirements_20231012004757.txt
captum==0.6.0
matplotlib==3.3.4
numpy==1.21.2
pandas==1.2.4
Pillow==10.0.0
scikit_learn==1.3.0
scipy==1.6.2
seaborn==0.11.1
timm==0.9.7
torch==1.10.0
torchvision==0.11.1+cu113
~imm==0.5.4
